fix: UI on list and edit create

+++ app/src/main/java/com/sawitpro/weightbridge/data/model/dto/WeighingTicketDto.kt
@@ -5,12 +5,16 @@ import com.squareup.moshi.Json
 data class WeighingTicketDto(
     @Json(name = "date")
     val date: String? = null,
+
     @Json(name = "driver_name")
     val driverName: String? = null,
+
     @Json(name = "license_number")
     val licenseNumber: String? = null,
+
     @Json(name = "inbound_weight")
-    val inboundWeight: Int? = null,
+    val inboundWeight: Double? = null,
+
     @Json(name = "outbound_weight")
-    val outboundWeight: Int? = null
+    val outboundWeight: Double? = null
 )
