+++ app/src/main/java/com/sawitpro/weightbridge/data/model/entity/UpdateWeighingTicketEntity.kt
@@ -4,7 +4,7 @@ data class UpdateWeighingTicketEntity(
     val driverName: String,
     val licenseNumber: String,
     val date: String,
-    val inboundWeight: Int,
-    val outboundWeight: Int,
-    val netWeight: Int
+    val inboundWeight: Double,
+    val outboundWeight: Double,
+    val netWeight: Double
 )
