+++ app/src/main/java/com/sawitpro/weightbridge/data/model/entity/WeighingTicketEntity.kt
@@ -10,7 +10,7 @@ data class WeighingTicketEntity(
     val driverName: String,
     val licenseNumber: String,
     val date: String,
-    val inboundWeight: Int,
-    val outboundWeight: Int,
-    val netWeight: Int
+    val inboundWeight: Double,
+    val outboundWeight: Double,
+    val netWeight: Double
 )
